LayoutTests/fast/page-color-sampling/color-sampling-fixed-popup.html

<!DOCTYPE html> <!-- webkit-test-runner [ ContentInsetBackgroundFillEnabled=true pageTopColorSamplingEnabled=true useFlexibleViewport=true ] -->
<html>
<head>
    <meta charset="utf-8">
    <meta name="viewport" content="width=device-width, initial-scale=1">
    <style>
        body, html {
            width: 100%;
            height: 100%;
            margin: 0;
            font-family: system-ui;
        }

        .popup {
            position: fixed;
            top: 0;
            bottom: 0;
            left: 0;
            right: 0;
        }

        .popup-content {
            background: rgb(100, 100, 100);
            width: 100%;
            height: 100%;
        }
    </style>
    <script src="../../resources/js-test.js"></script>
    <script src="../../resources/ui-helper.js"></script>
    <script>
    jsTestIsAsync = true;

    addEventListener("load", async () => {
        await UIHelper.setObscuredInsets(50, 50, 50, 50);
        await UIHelper.ensurePresentationUpdate();
        edgeColorsBeforeUnparenting = await UIHelper.fixedContainerEdgeColors();
        shouldBeEqualToString("edgeColorsBeforeUnparenting.top", "rgb(100, 100, 100)");
        shouldBeEqualToString("edgeColorsBeforeUnparenting.left", "rgb(100, 100, 100)");
        shouldBeEqualToString("edgeColorsBeforeUnparenting.right", "rgb(100, 100, 100)");
        shouldBeEqualToString("edgeColorsBeforeUnparenting.bottom", "rgb(100, 100, 100)");

        document.querySelector(".popup").remove();
        edgeColorsAfterUnparenting = await UIHelper.fixedContainerEdgeColors();
        shouldBeNull("edgeColorsAfterUnparenting.top");
        shouldBeNull("edgeColorsAfterUnparenting.left");
        shouldBeNull("edgeColorsAfterUnparenting.right");
        shouldBeNull("edgeColorsAfterUnparenting.bottom");

        finishJSTest();
    });
    </script>
</head>
<body>
<div class="popup">
    <div class="popup-content"></div>
</div>
<div class="tall"></div>
</body>
</html>
