LayoutTests/fast/page-color-sampling/color-sampling-fixed-popup-expected.txt

PASS edgeColorsBeforeUnparenting.top is "rgb(100, 100, 100)"
PASS edgeColorsBeforeUnparenting.left is "rgb(100, 100, 100)"
PASS edgeColorsBeforeUnparenting.right is "rgb(100, 100, 100)"
PASS edgeColorsBeforeUnparenting.bottom is "rgb(100, 100, 100)"
PASS edgeColorsAfterUnparenting.top is null
PASS edgeColorsAfterUnparenting.left is null
PASS edgeColorsAfterUnparenting.right is null
PASS edgeColorsAfterUnparenting.bottom is null
PASS successfullyParsed is true

TEST COMPLETE
